Extract the LLM subsystem into @corpus/core/llm
First slice of the shared corpus-agnostic engine. Moves the six LLM
modules — llm.ts, budget.ts, pricing.ts, ai-gateway.ts, settings.ts,
llm-error.ts — verbatim from packages/talmud/src/worker into
packages/core/src/llm. They form a self-contained subsystem (every import
is within the group; the only ambient types are KVNamespace/Ai from
workers-types), so the move is behaviour-preserving:

- Internal imports between the six stay relative and unchanged.
- 9 talmud importers + 5 LLM tests rewritten to @corpus/core/llm/<module>
  (1:1 path remap; no barrel, to avoid any LLMError re-export ambiguity).
- One vi.mock specifier updated to match the new import path.
- packages/core gains a typecheck script (tsc --noEmit) + its own tsconfig,
  so pnpm -r typecheck now compiles core standalone — proving it carries no
  Talmud dependency.

The model catalog + default chain (settings.ts) are LLM infra, not Talmud
content; each app still overrides the default per-deploy via the
DEFAULT_LLM_MODEL env var.

CLAUDE.md: document the workspace layout and that worktrees now want a real
pnpm install (workspace links for @corpus/core that vite/vitest need),
rather than symlinking node_modules from main.

Verified: pnpm -r typecheck, 1919 tests pass, vite build, wrangler
deploy --dry-run (4290 KiB worker, bindings intact). Codex read-only
review: no findings.
