+++ packages/core/src/llm/budget.ts
