+++ packages/core/src/llm/pricing.ts
