+++ packages/core/src/llm/settings.ts
